# docs/api_contract.yaml
openapi: 3.0.0
info:
  title: EmpPulse API
  version: 1.0.0
  description: REST API for EmpPulse employee leave, department, and time tracking.
servers:
   - url: http://localhost:8080
     description: Local
security:
  - CookieAuth: []
...
paths:
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Sign in and create a session
      description: Authenticate with email and password and receive a JSESSIONID cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session; returns the caller's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Sign out and invalidate the current session
      responses:
        '200':
          description: Session invalidated
          content:
            text/plain:
              schema:
                type: string
                example: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/password/forgot:
    post:
      tags:
        - Auth
      summary: Request a password reset email
      description: >-
        Sends a password reset email to the given address. Returns 404 if the address does not
        belong to a registered user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordForgotRequest'
      responses:
        '204':
          description: Reset email queued
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/password/reset:
    post:
      tags:
        - Auth
      summary: Reset password with token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: Password reset
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: >
        Shape returned by the auth/access-control handlers (a machine-readable code plus
        message). Domain/business-rule errors are emitted by Spring's default /error handler
        and carry only timestamp/status/error/path (no message body).
      properties:
        code:
          type: string
        message:
          type: string
    Theme:
      type: string
      enum:
        - light
        - dark
        - system
    Language:
      type: string
      enum:
        - en
        - ukr
    LeaveType:
      type: string
      enum:
        - vacation
        - sick
        - personal
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    PasswordForgotRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PasswordResetRequest:
      type: object
      required:
        - token
        - newPassword
        - confirmNewPassword
      properties:
        token:
          type: string
        newPassword:
          type: string
          format: password
        confirmNewPassword:
          type: string
          format: password
...
    UserPreferences:
      type: object
      required:
        - theme
        - language
      properties:
        theme:
          $ref: '#/components/schemas/Theme'
        language:
          $ref: '#/components/schemas/Language'
...
    User:
      type: object
      required:
        - id
        - name
        - surname
        - email
        - owner
        - preferences
      properties:
        id:
          type: integer
        name:
          type: string
        surname:
          type: string
        email:
          type: string
          format: email
        owner:
          type: boolean
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        employeeProfile:
          allOf:
            - $ref: '#/components/schemas/EmployeeProfile'
          nullable: true
        adminProfile:
          allOf:
            - $ref: '#/components/schemas/AdminProfile'
          nullable: true
    EmployeeProfile:
      type: object
      required:
        - employeeId
        - yearlyVacationBalance
        - vacationBalance
      properties:
        employeeId:
          type: integer
        departmentId:
          type: integer
          nullable: true
        departmentName:
          type: string
          nullable: true
        yearlyVacationBalance:
          type: integer
          description: The base annual allowance configured on the employee.
        vacationBalance:
          type: integer
          description: Days left, computed as yearlyVacationBalance + bonus days − used vacation days.
        activeLeave:
          allOf:
            - $ref: '#/components/schemas/ActiveLeave'
          nullable: true
    AdminProfile:
      type: object
      required:
        - adminId
        - departmentIds
      properties:
        adminId:
          type: integer
        departmentIds:
          type: array
          items:
            type: integer
    ActiveLeave:
      type: object
      required:
        - type
        - startDate
        - endDate
      properties:
        type:
          $ref: '#/components/schemas/LeaveType'
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...